Export all countries summaries to Excel

export_all_countries_summaries_xls Export summaries of all countries for the three billions or for a specific billion.

Usage

export_all_countries_summaries_xls(
  df,
  billion = c("hpop", "hep", "uhc", "all"),
  value_col = "value",
  scenario_col = NULL,
  transform_value_col = "transform_value",
  contribution = "contribution",
  contribution_pct = paste0(contribution, "_percent"),
  contribution_pct_total_pop = paste0(contribution, "_percent_total_pop"),
  default_scenario = "default_scenario",
  start_year = 2018,
  end_year = 2019:2025,
  scenario_end_year = end_year,
  output_folder = "outputs",
  version = whdh::get_formatted_timestamp()
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

billion

Billion indicator names to return, either "hep", "hpop", "uhc" , or "all".

value_col

Column name of column with indicator values.

scenario_col

Column name of column with scenario identifiers. Useful for calculating contributions on data in long format rather than wide format.

transform_value_col

Column name of column(s) with transformed indicator values, used to calculate contributions.

contribution

Column name of column(s) to store contribution (population) values. Must be the same length as transform_value_col.

contribution_pct

Column name of column(s) to store contribution (percent) values. Must be the same length as transform_value_col.

contribution_pct_total_pop

Column name of column(s) to store contribution (percent of total population of the country) values. Must be the same length as transform_value_col.

default_scenario

name of the default scenario.

start_year

Base year for contribution calculation, defaults to 2018.

end_year

End year(s) for contribution calculation, defaults to 2019 to 2025.

scenario_end_year

End year of the scenarios. Defaults to end_year.

output_folder

Folder path to where the Excel files should be written

version

character vector identifying the version of the plot to be passed.

Value

list of openxslx Workbook object. Output file is in output_folder.